flask-backend-mongo/api/requirements.txt
PyJWT
PyMySQL
Flask
python-decouple
gunicorn
flask-cors
pytest
requests
pymongo
